Change plasticity spike timing
Never report a pre-synaptic spike while replaying the spiking history of
the target neurons and always report a pre-synaptic spike once the
history is consumed, *potentially with n=0*.

Pre-synaptic spikes ought to be reported at time t+delay, where t is the
time point when the spike originated. However, for this, the synapse
must already be in state t+dt.