Implement CheckNeuron() function
constexpr bool CheckNeuron<T>() is meant to be used inside
static_assert()'s where T is a user-defined neuron. It checks that T
conforms to the Neuron concept and prints human-readable error messages
if it doesn't. In contrast to Neuron<T> it attempts to guess the user's
intention and suggest fixes.